requirements.txt
# Core dependencies
langchain>=0.1.0
langchain_community>=0.0.10
langchain_core>=0.1.10
langchain_chroma>=0.0.1
pydantic>=2.0.0

# LLM and embeddings
transformers>=4.35.0
sentence_transformers>=2.2.2
torch>=2.0.0
accelerate>=0.25.0
huggingface_hub>=0.19.0
bitsandbytes>=0.41.0  # For quantization support
llama-cpp-python>=0.2.0  # For GGUF model support

# Vector database
chromadb>=0.4.18

# Document processing
pypdf>=3.17.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
pyyaml>=6.0.1
unstructured>=0.10.0
markdown>=3.5.0

# Web frameworks
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.1
requests>=2.31.0
numpy>=1.24.0
